cookbook/pocoflow-agent/requirements.txt
pocoflow>=0.2.0
anthropic>=0.25.0
python-dotenv>=1.0
pyyaml>=6.0
duckduckgo-search>=7.5.2
